Added unit test and cleaned up some structure
